refactor agent actions
