+++ simulation_v2/agents/constants.py
@@ -1,0 +1,9 @@
+"""Simulation v2 agent action tuning constants."""
+
+PROB_LIKE_POST = 0.25
+PROB_WRITE_POST = 0.05
+PROB_FOLLOW_USER = 0.02
+
+MAX_POSTS_TO_LIKE_PER_TURN = 10
+MAX_POSTS_TO_WRITE_PER_TURN = 5
+MAX_USERS_TO_FOLLOW_PER_TURN = 5
